@@ -30,86 +30,86 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
-    "@ngneat/lib": "^3.0.1",
-    "@ngrx/effects": "^13.0.2",
-    "@ngrx/router-store": "^13.0.2",
-    "@ngrx/store": "^13.0.2",
-    "@reduxjs/toolkit": "^1.8.0",
-    "rxjs": "~7.4.0",
-    "tslib": "^2.3.1",
-    "zone.js": "~0.11.5"
+    "@angular/animations": "^14.2.2",
+    "@angular/common": "^14.2.2",
+    "@angular/compiler": "^14.2.2",
+    "@angular/core": "^14.2.2",
+    "@angular/forms": "^14.2.2",
+    "@angular/platform-browser": "^14.2.2",
+    "@angular/platform-browser-dynamic": "^14.2.2",
+    "@angular/router": "^14.2.2",
+    "@ngneat/lib": "^5.0.0",
+    "@ngrx/effects": "^14.3.1",
+    "@ngrx/router-store": "^14.3.1",
+    "@ngrx/store": "^14.3.1",
+    "@reduxjs/toolkit": "~1.8.5",
+    "rxjs": "~7.5.6",
+    "tslib": "^2.4.0",
+    "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^13.0.3",
-    "@angular-devkit/build-angular": "^13.3.0",
-    "@angular-eslint/builder": "^13.1.0",
-    "@angular-eslint/eslint-plugin": "^13.1.0",
-    "@angular-eslint/eslint-plugin-template": "^13.1.0",
-    "@angular-eslint/schematics": "^13.1.0",
-    "@angular-eslint/template-parser": "^13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@commitlint/cli": "^16.2.3",
-    "@commitlint/config-conventional": "^16.2.1",
-    "@ngrx/store-devtools": "^13.0.2",
+    "@angular-builders/jest": "^14.0.1",
+    "@angular-devkit/build-angular": "^14.2.3",
+    "@angular-eslint/builder": "^14.1.1",
+    "@angular-eslint/eslint-plugin": "^14.1.1",
+    "@angular-eslint/eslint-plugin-template": "^14.1.1",
+    "@angular-eslint/schematics": "^14.1.1",
+    "@angular-eslint/template-parser": "^14.1.1",
+    "@angular/cli": "^14.2.3",
+    "@angular/compiler-cli": "^14.2.2",
+    "@commitlint/cli": "^17.1.2",
+    "@commitlint/config-conventional": "^17.1.0",
+    "@ngrx/store-devtools": "^14.3.1",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
-    "@tailwindcss/forms": "^0.5.0",
-    "@testing-library/angular": "^11.0.3",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^12.20.47",
-    "@types/node-fetch": "^2.6.1",
-    "@typescript-eslint/eslint-plugin": "5.15.0",
-    "@typescript-eslint/parser": "5.15.0",
-    "all-contributors-cli": "^6.20.0",
-    "angular-cli-ghpages": "1.0.0",
+    "@tailwindcss/forms": "^0.5.3",
+    "@testing-library/angular": "^12.1.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.0.3",
+    "@types/node": "^12.20.55",
+    "@types/node-fetch": "^2.6.2",
+    "@typescript-eslint/eslint-plugin": "^5.38.0",
+    "@typescript-eslint/parser": "^5.38.0",
+    "all-contributors-cli": "^6.20.4",
+    "angular-cli-ghpages": "1.0.3",
     "cpx": "^1.5.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.11.0",
+    "eslint": "^8.23.1",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jest": "^26.1.1",
-    "eslint-plugin-jest-dom": "^4.0.1",
-    "eslint-plugin-jsdoc": "^38.0.4",
-    "eslint-plugin-ngrx": "^2.0.4",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.0.4",
+    "eslint-plugin-jest-dom": "^4.0.2",
+    "eslint-plugin-jsdoc": "^39.3.6",
+    "eslint-plugin-ngrx": "^2.1.4",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-rxjs": "^5.0.2",
-    "eslint-plugin-testing-library": "^5.1.0",
-    "git-cz": "^4.8.0",
-    "husky": "^7.0.4",
-    "jest": "^27.5.1",
-    "jest-preset-angular": "^11.1.1",
-    "jsonc-parser": "^3.0.0",
-    "lint-staged": "^12.3.6",
-    "msw": "^0.39.2",
-    "ng-packagr": "^13.2.1",
+    "eslint-plugin-testing-library": "^5.6.4",
+    "git-cz": "^4.9.0",
+    "husky": "^8.0.1",
+    "jest": "^29.0.3",
+    "jest-preset-angular": "^12.2.2",
+    "jsonc-parser": "^3.2.0",
+    "lint-staged": "^13.0.3",
+    "msw": "^0.47.3",
+    "ng-packagr": "^14.2.1",
     "node-fetch": "^2.6.7",
-    "postcss": "^8.4.12",
-    "prettier": "^2.6.0",
+    "postcss": "^8.4.16",
+    "prettier": "^2.7.1",
     "prettier-plugin-organize-attributes": "0.0.5",
     "protractor": "~7.0.0",
-    "semantic-release": "^18.0.1",
-    "stylelint": "^14.6.0",
+    "semantic-release": "^19.0.5",
+    "stylelint": "^14.12.0",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-recommended-scss": "^5.0.2",
-    "stylelint-config-standard": "^25.0.0",
-    "stylelint-config-standard-scss": "^3.0.0",
-    "stylelint-scss": "^4.2.0",
-    "tailwindcss": "^3.0.23",
-    "ts-jest": "^27.1.3",
-    "ts-node": "~9.1.1",
-    "typescript": "~4.4.4"
+    "stylelint-config-recommended-scss": "^7.0.0",
+    "stylelint-config-standard": "^28.0.0",
+    "stylelint-config-standard-scss": "^5.0.0",
+    "stylelint-scss": "^4.3.0",
+    "tailwindcss": "^3.1.8",
+    "ts-jest": "^29.0.1",
+    "ts-node": "~10.9.1",
+    "typescript": "~4.8.3"
   },
   "config": {
     "commitizen": {
